+++ app/Livewire/PembayaranTable.php
@@ -6,6 +6,8 @@ use App\Models\Pembayaran;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\Storage;
+use Masmerise\Toaster\Toastable;
 use PowerComponents\LivewirePowerGrid\Button;
 use PowerComponents\LivewirePowerGrid\Column;
 use PowerComponents\LivewirePowerGrid\Detail;
@@ -21,16 +23,14 @@ use PowerComponents\LivewirePowerGrid\Traits\WithExport;
 final class PembayaranTable extends PowerGridComponent
 {
     use WithExport;
+    use Toastable;
     public $transaksi_id;
 
     public function setUp(): array
     {
         $this->showCheckBox();
 
-        return [
-            Exportable::make('export')
-                ->striped()
-                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV),
+        $setUp = [
             Header::make()->showSearchInput(),
             Footer::make()
                 ->showPerPage()
@@ -39,23 +39,28 @@ final class PembayaranTable extends PowerGridComponent
                 ->showCollapseIcon()
                 ->view('details.pembayaran-detail'),
         ];
+
+        if (auth()->user()->can('export')) {
+            $setUp[] =
+                Exportable::make('export')
+                ->striped()
+                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV);
+        }
+
+        return $setUp;
     }
 
     public function datasource(): Builder
     {
         return Pembayaran::query()
-            ->leftJoin('users', 'pembayaran.user_id', '=', 'users.id')
-            ->leftJoin('rekening', 'pembayaran.rekening_id', '=', 'rekening.id')
-            ->leftJoin('transaksi', 'pembayaran.transaksi_id', '=', 'transaksi.id')
-            ->select('pembayaran.*', 'users.name as name', 'rekening.nama_bank as nama_bank', 'transaksi.total_harga as total_harga ', 'transaksi.status as status');
+            ->with('user', 'rekening', 'transaksi');
     }
 
     public function relationSearch(): array
     {
         return [
-            'users.name' => 'name',
-            'rekening' => 'nama_bank',
-            'transaksi' => ['total_harga', 'status'],
+            'user' => ['name'],
+            'rekening' => ['nama_rekening'],
         ];
     }
 
@@ -66,10 +71,12 @@ final class PembayaranTable extends PowerGridComponent
             ->add('transaksi_id')
             ->add('user_id')
             ->add('rekening_id')
-            ->add('name')
+            ->add('user.name')
             ->add('foto')
-            ->add('total_harga')
-            ->add('status_transaksi')
+            ->add('transaksi.total_harga', function ($row) {
+                return 'Rp ' . number_format($row->transaksi->total_harga, 0, ',', '.');
+            })
+            ->add('status')
             ->add('nama_rekening');
     }
 
@@ -80,32 +87,17 @@ final class PembayaranTable extends PowerGridComponent
                 ->sortable()
                 ->searchable(),
 
-            Column::make('Nama ', 'name')
+            Column::make('Nama ', 'user.name')
                 ->sortable()
                 ->searchable(),
 
-            Column::make('Nama ', 'transaksi_id')
+            Column::make('Total Harga ', 'transaksi.total_harga')
                 ->sortable()
                 ->searchable(),
 
-            // Column::make('Foto', 'foto')
-            //     ->sortable()
-            //     ->searchable(),
-
-            // Column::make('Total', 'total_harga')
-            //     ->sortable()
-            //     ->searchable(),
-
-            // Column::make('Nama rekening', 'nama_rekening')
-            //     ->sortable()
-            //     ->searchable(),
-
-            // Column::make('Status', 'status')
-            //     ->sortable()
-            //     ->searchable(),
-
-            // Column::make('Created at', 'created_at_formatted', 'created_at')
-            //     ->sortable(),
+            Column::make('Status', 'status')
+                ->sortable()
+                ->searchable(),
 
             Column::action('Action')
         ];
@@ -116,52 +108,54 @@ final class PembayaranTable extends PowerGridComponent
         return [];
     }
 
-    #[\Livewire\Attributes\On('edit')]
-    public function edit($rowId): void
-    {
-        $this->js('alert(' . $rowId . ')');
-    }
-
     public function actions(\App\Models\Pembayaran $row): array
     {
-        return [
-            Button::add('edit')
+        $actions = [];
+
+        // Jika user memiliki permissions update
+        if (auth()->user()->can('verifikasi')) {
+            $actions[] =
+                Button::add('edit')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" class="w-5 h-5">
                 <path fill-rule="evenodd" d="M8.603 3.799A4.49 4.49 0 0 1 12 2.25c1.357 0 2.573.6 3.397 1.549a4.49 4.49 0 0 1 3.498 1.307 4.491 4.491 0 0 1 1.307 3.497A4.49 4.49 0 0 1 21.75 12a4.49 4.49 0 0 1-1.549 3.397 4.491 4.491 0 0 1-1.307 3.497 4.491 4.491 0 0 1-3.497 1.307A4.49 4.49 0 0 1 12 21.75a4.49 4.49 0 0 1-3.397-1.549 4.49 4.49 0 0 1-3.498-1.306 4.491 4.491 0 0 1-1.307-3.498A4.49 4.49 0 0 1 2.25 12c0-1.357.6-2.573 1.549-3.397a4.49 4.49 0 0 1 1.307-3.497 4.49 4.49 0 0 1 3.497-1.307Zm7.007 6.387a.75.75 0 1 0-1.22-.872l-3.236 4.53L9.53 12.22a.75.75 0 0 0-1.06 1.06l2.25 2.25a.75.75 0 0 0 1.14-.094l3.75-5.25Z" clip-rule="evenodd" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->openModal('pembayaran-form', ['rowId' => $row->id, 'updatingStatusOnly' => false, 'transaksi_id' => $row->transaksi_id]),
-            Button::add('delete')
+                ->openModal('pembayaran-form', ['rowId' => $row->id, 'updatingStatusOnly' => true]);
+        }
+
+        // Jika user memiliki permissions delete
+        if (auth()->user()->can('delete')) {
+            $actions[] =
+                Button::add('delete')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->dispatch('delete', ['rowId' => $row->id])
-        ];
+                ->dispatch('delete', ['rowId' => $row->id]);
+        }
+        return $actions;
     }
 
     public function header(): array
     {
-        return [
-            Button::add('add-pembayaran')
-                ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
-                <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
-                </svg>
-                '))
-                ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->openModal('pembayaran-form', ['transaksi_id' => $this->transaksi_id, 'updatingStatusOnly' => false]),
-            Button::add('export-pdf')
+        $header = [];
+
+        // Jika user memiliki permissions export
+        if (auth()->user()->can('export')) {
+            $header[] =
+                Button::add('export-pdf')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0 1 10.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0 .229 2.523a1.125 1.125 0 0 1-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0 0 21 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 0 0-1.913-.247M6.34 18H5.25A2.25 2.25 0 0 1 3 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 0 1 1.913-.247m10.5 0a48.536 48.536 0 0 0-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5Zm-3 0h.008v.008H15V10.5Z" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->dispatch('exportPdf', []),
-        ];
+                ->dispatch('exportPdf', []);
+        }
+        return $header;
     }
 
     protected function getListeners()
@@ -198,8 +192,11 @@ final class PembayaranTable extends PowerGridComponent
     public function delete($rowId)
     {
         $pembayaran = Pembayaran::findOrFail($rowId);
-        // Detach all associated users
-        $pembayaran->user()->detach();
+        if ($pembayaran->foto) {
+            Storage::disk('public')->delete($pembayaran->foto);
+        }
         $pembayaran->delete();
+
+        $this->success('Pembayaran berhasil dihapus');
     }
 }
